Rename Formating to Formatter.
