+++ core/src/main/java/com/novoda/noplayer/internal/mediaplayer/ErrorFactory.java
@@ -6,7 +6,7 @@ import com.novoda.noplayer.NoPlayerError;
 import com.novoda.noplayer.Player;
 import com.novoda.noplayer.PlayerErrorType;
 
-import static com.novoda.noplayer.internal.mediaplayer.ErrorFormating.formatMessage;
+import static com.novoda.noplayer.internal.mediaplayer.ErrorFormatter.formatMessage;
 
 public final class ErrorFactory {
