+++ core/src/test/java/com/novoda/noplayer/internal/mediaplayer/ErrorFormatterTest.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 
 import static org.fest.assertions.api.Assertions.assertThat;
 
-public class ErrorFormatingTest {
+public class ErrorFormatterTest {
 
     private static final int TYPE_CODE = 202;
     private static final int EXTRA_CODE = -218;
@@ -13,7 +13,7 @@ public class ErrorFormatingTest {
     public void givenTypeAndExtra_whenFormattingMessage_thenReturnsExpectedMessageFormat() {
         String expectedFormat = "Type: 202, Extra: -218";
 
-        String actualFormat = ErrorFormating.formatMessage(TYPE_CODE, EXTRA_CODE);
+        String actualFormat = ErrorFormatter.formatMessage(TYPE_CODE, EXTRA_CODE);
 
         assertThat(actualFormat).isEqualTo(expectedFormat);
     }
